+++ examples/iced/test/rekindle_client_test.exs
@@ -1,0 +1,9 @@
+defmodule IcedExample.RekindleClientTest do
+  use ExUnit.Case, async: false
+
+  @tag :rust
+  @tag timeout: :infinity
+  test "Rust client" do
+    Rekindle.Test.run!(:iced_example)
+  end
+end