+++ examples/slint/test/rekindle_client_test.exs
@@ -1,0 +1,9 @@
+defmodule SlintExample.RekindleClientTest do
+  use ExUnit.Case, async: false
+
+  @tag :rust
+  @tag timeout: :infinity
+  test "Rust client" do
+    Rekindle.Test.run!(:slint_example)
+  end
+end